Cambio de campo para comparar las amenidades con las reservaciones

+++ database/migrations/2026_03_17_232832_update_reservations_table.php
@@ -2,36 +2,48 @@
 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
-use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
-        Schema::table('reservations.reservations', function (Blueprint $table) {
-            $table->foreignId('amenity_resource_id')->nullable();
-            $table->foreign('amenity_resource_id')->references('id')->on('amenity_resources');
-            $table->unsignedBigInteger('reservation_status_id')->nullable();
-            $table->foreign('reservation_status_id')->references('id')->on('reservations.status');
-            $table->index(['amenity_resource_id', 'reservation_date'], 'amenity_resource_id_reservation_date_index');
+        Schema::table('reservations', function (Blueprint $table) {
+
+            if (!Schema::hasColumn('reservations', 'amenity_resource_id')) {
+
+                $table->foreignId('amenity_resource_id')->nullable();
+
+                $table->foreign('amenity_resource_id')
+                    ->references('id')
+                    ->on('amenity_resources');
+
+            }
+
         });
+
+        DB::statement("
+            CREATE INDEX IF NOT EXISTS amenity_resource_id_date_index
+            ON reservations (amenity_resource_id, date)
+        ");
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
-        Schema::table('reservations.reservations', function (Blueprint $table) {
-            $table->dropForeign(['amenity_resource_id']);
-            $table->dropColumn('amenity_resource_id');
-            $table->dropForeign(['reservation_status_id']);
-            $table->dropColumn('reservation_status_id');
-            DB::statement('DROP INDEX IF EXISTS reservations.amenity_resource_id_reservation_date_index');
+        Schema::table('reservations', function (Blueprint $table) {
+
+            if (Schema::hasColumn('reservations', 'amenity_resource_id')) {
+
+                $table->dropForeign(['amenity_resource_id']);
+                $table->dropColumn('amenity_resource_id');
+
+            }
+
         });
+
+        DB::statement("
+            DROP INDEX IF EXISTS amenity_resource_id_date_index
+        ");
     }
 };
